Fixed baddies when creating and deleting

+++ src/main/java/com/revature/assignforce/service/FocusServiceImpl.java
@@ -1,7 +1,9 @@
 package com.revature.assignforce.service;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 import javax.transaction.Transactional;
 
@@ -9,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.revature.assignforce.beans.Focus;
+import com.revature.assignforce.beans.SkillIdHolder;
 import com.revature.assignforce.repos.FocusRepository;
 
 @Transactional
@@ -17,7 +20,7 @@ public class FocusServiceImpl implements FocusService {
 
 	@Autowired
 	private FocusRepository focusRepository;
-	
+
 	@Override
 	public List<Focus> getAll() {
 		return focusRepository.findAll();
@@ -35,11 +38,25 @@ public class FocusServiceImpl implements FocusService {
 
 	@Override
 	public Focus create(Focus b) {
+		Set<SkillIdHolder> skills = b.getSkills();
+		if (skills == null) {
+			skills = new HashSet<SkillIdHolder>();
+		}
+		b.setSkills(new HashSet<SkillIdHolder>());
+		focusRepository.save(b);
+		b.setSkills(skills);
+		focusRepository.flush();
 		return focusRepository.save(b);
 	}
 
 	@Override
 	public void delete(int id) {
+		Optional<Focus> focus = focusRepository.findById(id);
+		if (!focus.isPresent()) {
+			return;
+		}
+		focus.get().setSkills(new HashSet<SkillIdHolder>());
+		focusRepository.save(focus.get());
 		focusRepository.deleteById(id);
 	}
